requirements.txt
aniso8601==8.0.0
attrs==20.2.0
click==7.1.2
flasgger==0.9.5
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
marshmallow==3.8.0
mistune==0.8.4
pyrsistent==0.17.3
python-decouple==3.3
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
SQLAlchemy==1.3.19
Werkzeug==1.0.1
